+++ src/states/entity/Stove.lua
@@ -22,7 +22,19 @@ function Stove:init()
     self.duration = levelPerformance[self.level].duration
     self.stockType = 'Meat'
     self.stock = StockManager:getStockTotal()[self.stockType]
-    self._bubbleColor = gColors['purple']
+    self.bubbleColor = gColors['purple']
+    self.bubble = Bubble({
+        x = self.x,
+        y = self.y,
+        desired_width = self.desired_width,
+        desired_height = self.desired_height,
+        bubbleColor = self.bubbleColor,
+    })
+
+    self.animationCounter = 0
+
+    self.isAvail = false
+    gStateStack:push(self.bubble)
 end
 
 function Stove:update(dt)
@@ -32,15 +44,39 @@ function Stove:update(dt)
         if self.counter >= self.duration then
             self.counter = 0
             self.productionStage = 'Ready'
+            self.isAvail = true
+        end
+    end
+
+    if self.isAvail and self.productionStage ~= 'Holding' then
+        self.bubble:activate()
+    else
+        self.bubble:deactivate()
+    end
+
+    self:updateFrame()
+end
+
+
+function Stove:updateFrame()
+    if self.isAvail then
+        self.frame = self.frames[12]
+    elseif self.productionStage ~= 'Producing' then
+        self.frame = self.frames[1]
+    else
+        local frameDuration = self.duration / 10
+        for i = 2, 11 do
+            if self.counter > frameDuration * i then
+                self.frame = self.frames[i]
+            end
         end
     end
 end
 
 function Stove:render()
-    love.graphics.setColor(gColors['purple'])
-    love.graphics.rectangle('fill', self.x, self.y, self.desired_width, self.desired_height)
-    love.graphics.setColor(gColors['white'])
-    love.graphics.printf(self.type, self.x, self.y + 20, self.desired_width, 'center')
+    if self.frame then
+        love.graphics.draw(self.texture, self.frame, self.x, self.y)
+    end
 end
 
 function Stove:produce()
@@ -48,30 +84,21 @@ function Stove:produce()
         self.productionStage = 'Producing'
         self.counter = 0
         self.stock = StockManager:consume(self.stockType)
-        self:showBubble()
     end
 end
 
 function Stove:drag()
     self.productionStage = 'Holding'
-    --self:updateFrame()
-    self:hideBubble()
+    self:updateFrame()
 end
 
 function Stove:taken()
     self.productionStage = 'Ready'
-    --[[self:updateFrame()
-    if self.volume > 0 then
-        self:showBubble(self._bubbleColor)
-    else
-        self:hideBubble()
-    end]]
+    self:updateFrame()
+    self.isAvail = false
 end
 
 function Stove:undrag()
     self.productionStage = 'Ready'
-    --[[self:updateFrame()
-    if self.volume > 0 then
-        self:showBubble(self._bubbleColor)
-    end]]
+    self:updateFrame()
 end